# graph/path: replace dominators implementation