@kortschak 2017-09-01
This gives an interesting failure. The change of dfs from the recursive routine to the iterative results in about 9% failure. Presumably I have failed to note something during the `traverse.DepthFirst` walk.