@kortschak 2017-09-01
For a great bit of history, and IMO better figures, see http://www.dtic.mil/dtic/tr/fulltext/u2/a054144.pdf.