@mewmew 2017-09-05
Wow. Thanks for working on this @kortschak! As a user of the dominator tree calculation API, I see tremendous increase in performance when updating to use the Lengauer and Tarjan algorithm, as implemented by this PR.

Below follows a comparison of run time for the [restructure tool](https://godoc.org/github.com/decomp/decomp/cmd/restructure), when using the old and new algorithm:

# Before

## Graph with 32 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n32_graphs> time restructure main.dot 

real 0.32
user 0.32
sys 0.01
```

## Graph with 64 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n64_graphs> time restructure main.dot 

real 3.97
user 4.41
sys 0.04
```

## Graph with 128 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n128_graphs> time restructure main.dot 

real 71.65
user 89.00
sys 0.68
```

## Graph with 256 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n256_graphs> time restructure main.dot

Aborted after 2.5 minutes.
```

# After

## Graph with 64 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n64_graphs> time restructure main.dot 

real 0.03
user 0.02
sys 0.00
```

## Graph with 128 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n128_graphs> time restructure main.dot 

real 0.11
user 0.11
sys 0.00
```

## Graph with 256 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n256_graphs> time restructure main.dot
real 0.41
user 0.42
sys 0.00

## Graph with 512 nodes
```

~/D/g/s/g/d/t/p/t/if_n512_graphs> time restructure main.dot

real 1.79
user 1.91
sys 0.01

## Graph with 1024 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n1024_graphs> time restructure main.dot

real 8.39
user 9.81
sys 0.04
```

## Graph with 2048 nodes

```
u@x220 ~/D/g/s/g/d/t/p/t/if_n2048_graphs> time restructure main.dot 

real 33.77
user 39.55
sys 0.16
```