@mewmew 2017-09-05
Also, thanks for the reference to the paper. It is a happy read.