@mewmew 2017-09-05
As the API is refined and extended with this PR, I would like to propose adding a `Dominates` method on DominatorTree, defined as follows.

```go
// Dominates reports whether a immediately dominates b.
func (d DominatorTree) Dominates(a, b graph.Node) bool {
	return a == d.DominatorOf(b)
}
```

If however the wish is to keep the API minimal, then it is trivial for users to define this method themselves. This is what is done at https://github.com/decomp/decomp/blob/master/graph/cfg/dom.go

This was the old wrapper for implementing `Dominates`: https://github.com/decomp/decomp/blob/75933d7d4040fa551bcf608a7c76b073f378cc62/graph/cfg/dom.go