@kortschak 2017-09-05
I'm glad there is a set of data to benchmark performance. Are those graphs available somewhere? I'm thinking about adding in the sophisticated parts of the algorithm, but I would like to see the impact of that.

I won't add the Dominates method since I think it actually reduced readability; `a == d.DominatorOf(b)` is clearer to me than `d.Dominates(a, b)` since the meaning of `a` and  `b` is not obvious from the call without reading the documentation.

Also, you should note that the two implementations are different, so your old `Dominates` will give different answers than the new `Dominates`; the previous `path.Dominators` function returned all domination relationships for each node, while the new one only returns immediate domination relationships. So you should have the following if you want to match the old behaviour.

```
func dominates(t path.DoninatorTree, a, b graph.Node) bool {
    for b = t.DominatorOf(b) /* omit if including self domination */; b != nil; b = t.DominatorOf(b) {
        if b.ID() = a.ID() {
            return true
        }
    }
    return false
}
```