@mewmew 2017-09-05
> I'm glad there is a set of data to benchmark performance. Are those graphs available somewhere? I'm thinking about adding in the sophisticated parts of the algorithm, but I would like to see the impact of that.

Those graphs were automatically generated using [gen_nested](https://github.com/decomp/testdata/blob/master/_scripts_/gen_nested.go), which generates C code containing nested control flow primitives. The C source code was then converted to LLVM IR using Clang, and control flow graphs were created from each function using [ll2dot](https://godoc.org/github.com/decomp/decomp/cmd/ll2dot). Note, the `opt` tool of LLVM may also be used to create the control flow graphs (when passed the `-dot-cfg-only` command line flag).

Below follows a summary of the steps used to generate the CFGs.
```
git clone https://github.com/decomp/testdata
# Requires "flow" branch of gonum.
go get github.com/decomp/decomp/cmd/ll2dot
cd testdata/pathological/testdata
ll2dot *.ll
```

For other test data, you may also find the LLVM IR assembly of the tools included in the GNU Core Utilities and SQLite projects. (I don't think `ll2dot` is up for the task of parsing those LLVM IR files just, as DWARF debug data is yet to be supported). However, you could use `opt -dot-cfg-only -S foo.ll` to convert those files into CFGs.

Note, the `ll2dot` tool makes the entry/root node of functions with the DOT label attribute `entry`, so you can programatically parse each DOT file and generate the dominator tree based on the root node.

> I won't add the Dominates method since I think it actually reduced readability; a == d.DominatorOf(b) is clearer to me than d.Dominates(a, b) since the meaning of a and b is not obvious from the call without reading the documentation.

Fair enough, I understand your rationale. I will give it some pondering and potentially update on my end as well.

> Also, you should note that the two implementations are different, so your old Dominates will give different answers than the new Dominates; the previous path.Dominators function returned all domination relationships for each node, while the new one only returns immediate domination relationships.

Thanks for pointing this out. I had noticed the difference. Still have to do some further evaluation of which parts of the code that relies on this, and if they may be adapted to use immediate dominance information, or if the function needs to recurse to the root node.