@mewmew 2017-09-05
Note, the results from the CFGs of GNU Core Utilities is probably a much more reliable bench mark than the ones from `gen_nested`, as such code will never occur in the real world.