@kortschak 2017-09-05
Flow graph domination is not just valuable in CS. They can be used for analysis of biological network flows.