@mewmew 2017-09-05
> Flow graph domination is not just valuable in CS. They can be used for analysis of biological network flows.

That's true, I should not make assumptions on the data it will be used for. Rather, we should try to evaluate it on different categories of data; such as CFGs and biological network flows.