@kortschak 2017-09-06
I'll start with your nested if graphs and make a general framework that we can just drop dot files into.