@kortschak 2017-09-06
I've implemented the sophisticated algorithm, but I see the following results:

```
$ benchstat old.bench new.bench 
name                          old time/op  new time/op  delta
Dominators/nested_if_n1024-8  4.35ms ± 8%  4.37ms ± 6%    ~     (p=0.365 n=20+19)
Dominators/nested_if_n128-8    528µs ± 6%   528µs ± 4%    ~     (p=0.517 n=20+17)
Dominators/nested_if_n16-8    64.3µs ± 3%  65.5µs ± 2%  +1.73%  (p=0.000 n=20+19)
Dominators/nested_if_n2048-8  8.54ms ± 3%  8.71ms ± 6%  +2.03%  (p=0.005 n=20+18)
Dominators/nested_if_n256-8   1.09ms ± 7%  1.09ms ± 6%    ~     (p=0.947 n=20+20)
Dominators/nested_if_n32-8     128µs ± 4%   132µs ± 7%  +3.31%  (p=0.000 n=19+20)
Dominators/nested_if_n512-8   2.14ms ± 6%  2.20ms ± 5%  +2.67%  (p=0.003 n=19+19)
Dominators/nested_if_n64-8     257µs ± 6%   275µs ± 6%  +7.11%  (p=0.000 n=20+20)
```

So I will not push that. The diff follows in case we review this and decide otherwise.

```diff
diff --git a/graph/path/control_flow.go b/graph/path/control_flow.go
index 3c472a8..b1ced69 100644
--- a/graph/path/control_flow.go
+++ b/graph/path/control_flow.go
@@ -9,12 +9,15 @@ import "gonum.org/v1/gonum/graph"
 // Dominators returns a dominator tree for all nodes in the flow graph
 // g starting from the given root node.
 func Dominators(root graph.Node, g graph.Directed) DominatorTree {
-       // The algorithm used here is essentially the Lengauer and Tarjan
-       // algorithm described in https://doi.org/10.1145%2F357062.357071
+       // The algorithm used here is essentially the sophisticated
+       // Lengauer and Tarjan algorithm described in
+       // https://doi.org/10.1145%2F357062.357071
 
        lt := lengauerTarjan{
                indexOf: make(map[int64]int),
+               base:    ltNode{semi: -1},
        }
+       lt.base.label = &lt.base
 
        // step 1.
        lt.dfs(g, root)
@@ -80,6 +83,12 @@ type lengauerTarjan struct {
        // number of the node in the Lengauer-
        // Tarjan algorithm.
        indexOf map[int64]int
+
+       // base is the base label for balanced
+       // tree path compression used in the
+       // sophisticated Lengauer-Tarjan
+       // algorith,
+       base ltNode
 }
 
 // ltNode is a graph node with accounting for the Lengauer-Tarjan
@@ -104,6 +113,14 @@ type ltNode struct {
        //      is the number of the semidominator of w.
        semi int
 
+       // size is the tree size of w used in the
+       // sophisticated algorithm.
+       size int
+
+       // child is the child node of w used in the
+       // sophisticated algorithm.
+       child *ltNode
+
        // bucket is the set of vertices whose
        // semidominator is w.
        bucket map[*ltNode]struct{}
@@ -135,6 +152,8 @@ func (lt *lengauerTarjan) dfs(g graph.Directed, v graph.Node) {
        ltv := &ltNode{
                node:   v,
                semi:   i,
+               size:   1,
+               child:  &lt.base,
                bucket: make(map[*ltNode]struct{}),
        }
        ltv.label = ltv
@@ -176,15 +195,37 @@ func (lt *lengauerTarjan) compress(v *ltNode) {
 // eval is the Lengauer-Tarjan EVAL function.
 func (lt *lengauerTarjan) eval(v *ltNode) *ltNode {
        if v.ancestor == nil {
-               return v
+               return v.label
        }
        lt.compress(v)
-       return v.label
+       if v.ancestor.label.semi >= v.label.semi {
+               return v.label
+       }
+       return v.ancestor.label
 }
 
 // link is the Lengauer-Tarjan LINK procedure.
 func (*lengauerTarjan) link(v, w *ltNode) {
-       w.ancestor = v
+       s := w
+       for w.label.semi < s.child.label.semi {
+               if s.size+s.child.child.size >= 2*s.child.size {
+                       s.child.ancestor = s
+                       s.child = s.child.child
+               } else {
+                       s.child.size = s.size
+                       s.ancestor = s.child
+                       s = s.child
+               }
+       }
+       s.label = w.label
+       v.size += w.size
+       if v.size < 2*w.size {
+               s, v.child = v.child, s
+       }
+       for s != nil {
+               s.ancestor = v
+               s = s.child
+       }
 }
 
 // DominatorTree is a flow graph dominator tree.
```