@kortschak 2017-09-07
I have found a set of cases where the sophisticated algorithm is an improvement; the Erdős-Rényi model graph G(n,m) shows modest to quite significant improvements.

```
$ benchstat simple.bench sophisticated.bench 
name                            old time/op  new time/op  delta
RandomDominators/gnm-1e3-1e4-8  3.64ms ± 2%  3.56ms ± 4%  -2.20%  (p=0.008 n=9+10)
RandomDominators/gnm-1e4-1e5-8  47.4ms ± 2%  46.2ms ± 2%  -2.59%  (p=0.000 n=10+10)
RandomDominators/gnm-1e5-1e6-8   790ms ± 1%   759ms ± 2%  -3.92%  (p=0.000 n=9+10)
RandomDominators/gnm-1e6-1e7-8   10.6s ± 4%    9.6s ± 4%  -9.22%  (p=0.000 n=10+10)
```

So now we are in a situation were there is argument for both. Maybe we should do that; include the current `Dominators` function and also have `DominatorsSophisticated` (name suggestions welcomed).

I'll wait for review before I go ahead with that.

Benchmark code:

```
// +build go1.7

package path

import (
	"math/rand"
	"testing"

	"gonum.org/v1/gonum/graph"
	"gonum.org/v1/gonum/graph/graphs/gen"
	"gonum.org/v1/gonum/graph/simple"
	"gonum.org/v1/gonum/graph/topo"
)

func BenchmarkRandomDominators(b *testing.B) {
	tests := []struct {
		name string
		g    graph.Directed
	}{
		{name: "gnm-1e3-1e4", g: gnm(1e3, 1e4)},
		{name: "gnm-1e4-1e5", g: gnm(1e4, 1e5)},
		{name: "gnm-1e5-1e6", g: gnm(1e5, 1e6)},
		{name: "gnm-1e6-1e7", g: gnm(1e6, 1e7)},
	}

	for _, test := range tests {
		sort, err := topo.Sort(test.g)
		root := sort[0]
		if root == nil {
			// If we did not get a node in the first position
			// then there must be an unorderable set of node
			// in the first position of the error. Pick one
			// at random.
			root = err.(topo.Unorderable)[0][0]
		}
		if root == nil {
			b.Error("no entry node label for graph")
			continue
		}

		b.Run(test.name, func(b *testing.B) {
			for i := 0; i < b.N; i++ {
				d := Dominators(root, test.g)
				if got := d.Root(); got.ID() != root.ID() {
					b.Fatalf("unexpected root node: got:%d want:%d", got.ID(), root.ID())
				}
			}
		})
	}
}

func gnm(n, m int) graph.Directed {
	dg := simple.NewDirectedGraph()
	err := gen.Gnm(dg, n, m, rand.New(rand.NewSource(1)))
	if err != nil {
		panic(err)
	}
	return dg
}
```